modularized amortized model
